+++ Gw2Sharp/WebApi/V2/Clients/Characters/ICharactersIdClient.cs
@@ -1,9 +1,13 @@
+using Gw2Sharp.WebApi.V2.Models;
+
 namespace Gw2Sharp.WebApi.V2.Clients
 {
     /// <summary>
     /// A client of the Guild Wars 2 API v2 characters id endpoint.
     /// </summary>
-    public interface ICharactersIdClient : IClient
+    public interface ICharactersIdClient :
+        IAuthenticatedClient<Character>,
+        IBlobClient<Character>
     {
         /// <summary>
         /// The character name.